msa-backing-service/auth_service/app/core/security.py
from datetime import datetime,timedelta,timezone
from typing import Optional,Dict,Any,List
from jose import JWTError,jwt
from passlib.context import CryptContext
from app.core.config import settings
import uuid
pwd_context=CryptContext(schemes=['bcrypt'],deprecated='auto')
def verify_password(plain_password,hashed_password):return pwd_context.verify(plain_password,hashed_password)
def get_password_hash(password):return pwd_context.hash(password)
def create_access_token(data,expires_delta=None):
	A=expires_delta;B=data.copy()
	if A:C=datetime.now(timezone.utc)+A
	else:C=datetime.now(timezone.utc)+timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
	B.update({'exp':C,'iat':datetime.now(timezone.utc),'jti':str(uuid.uuid4())});D=jwt.encode(B,settings.SECRET_KEY,algorithm=settings.ALGORITHM);return D
def create_refresh_token(data,expires_delta=None):
	A=expires_delta;B=data.copy()
	if A:C=datetime.now(timezone.utc)+A
	else:C=datetime.now(timezone.utc)+timedelta(days=settings.REFRESH_TOKEN_EXPIRE_DAYS)
	D=str(uuid.uuid4());B.update({'exp':C,'iat':datetime.now(timezone.utc),'jti':D,'type':'refresh'});E=jwt.encode(B,settings.SECRET_KEY,algorithm=settings.ALGORITHM);return E,D
def decode_token(token):
	try:A=jwt.decode(token,settings.SECRET_KEY,algorithms=[settings.ALGORITHM]);return A
	except JWTError:return