msa-backing-service/auth_service/requirements.txt
fastapi[all]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator